Move plugins recorder streamer and tasks into main NDS namespace
