+++ src/neural_data_simulator/plugins/examples/custom_script.py
