+++ src/neural_data_simulator/plugins/examples/model.py
