+++ src/neural_data_simulator/plugins/examples/postprocessor.py
