+++ src/neural_data_simulator/plugins/examples/tests/test_model.py
