+++ src/neural_data_simulator/scripts/post_install_config.py
@@ -13,13 +13,13 @@ import os
 import shutil
 from urllib.parse import urljoin
 
-import decoder
-import plugins
 import pooch
-import streamer
-import tasks
 
 import neural_data_simulator.core as core
+from neural_data_simulator import decoder
+from neural_data_simulator import plugins
+from neural_data_simulator import streamer
+from neural_data_simulator import tasks
 from neural_data_simulator.util.runtime import get_abs_path
 from neural_data_simulator.util.runtime import get_configs_dir
 from neural_data_simulator.util.runtime import get_plugins_dir
