+++ src/neural_data_simulator/streamer/config/settings_streamer.yaml
