+++ src/neural_data_simulator/streamer/run_streamer.py
@@ -19,13 +19,13 @@ from typing import cast, Dict, Iterator, List, Optional, Tuple
 from neo.rawio.blackrockrawio import BlackrockRawIO
 import numpy as np
 from pydantic_yaml import VersionedYamlModel
-from streamer import settings
-from streamer import streamers
 
 from neural_data_simulator.core.outputs import LSLOutputDevice
 from neural_data_simulator.core.outputs import StreamConfig
 from neural_data_simulator.core.samples import Samples
 from neural_data_simulator.core.settings import LogLevel
+from neural_data_simulator.streamer import settings
+from neural_data_simulator.streamer import streamers
 from neural_data_simulator.util.runtime import configure_logger
 from neural_data_simulator.util.runtime import get_abs_path
 from neural_data_simulator.util.runtime import initialize_logger
