+++ src/neural_data_simulator/streamer/settings.py
