+++ src/neural_data_simulator/streamer/streamers.py
