+++ src/neural_data_simulator/tasks/center_out_reach/joystick.py
