+++ src/neural_data_simulator/tasks/center_out_reach/screen_info.py
