+++ src/neural_data_simulator/tasks/config/settings_center_out_reach.yaml
