+++ src/neural_data_simulator/tasks/run_center_out_reach.py
@@ -7,15 +7,15 @@ from typing import cast, Tuple
 
 import numpy as np
 from pydantic_yaml import VersionedYamlModel
-from tasks.center_out_reach.input_events import InputHandler
-from tasks.center_out_reach.metrics import MetricsCollector
-from tasks.center_out_reach.scalers import PixelsToMetersConverter
-from tasks.center_out_reach.scalers import StandardVelocityScaler
-from tasks.center_out_reach.settings import CenterOutReach
-from tasks.center_out_reach.task_runner import TaskRunner
-from tasks.center_out_reach.task_state import StateParams
-from tasks.center_out_reach.task_state import TaskState
-from tasks.center_out_reach.task_window import TaskWindow
+from neural_data_simulator.tasks.center_out_reach.input_events import InputHandler
+from neural_data_simulator.tasks.center_out_reach.metrics import MetricsCollector
+from neural_data_simulator.tasks.center_out_reach.scalers import PixelsToMetersConverter
+from neural_data_simulator.tasks.center_out_reach.scalers import StandardVelocityScaler
+from neural_data_simulator.tasks.center_out_reach.settings import CenterOutReach
+from neural_data_simulator.tasks.center_out_reach.task_runner import TaskRunner
+from neural_data_simulator.tasks.center_out_reach.task_state import StateParams
+from neural_data_simulator.tasks.center_out_reach.task_state import TaskState
+from neural_data_simulator.tasks.center_out_reach.task_window import TaskWindow
 
 from neural_data_simulator.core import inputs
 from neural_data_simulator.core import outputs
